+++ src/dev/theskidster/mapeditor/main/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains files essential in the applications execution.
+ */
+package dev.theskidster.mapeditor.main;